move code to more appropriate place
